Add SqStrArray files. They are separated from the SqPtrArray files.
For C language:
SqStrArray is changed from SqStringArray. It copy string when appending and inserting, which is different from SqStringArray of old version.

For C++ language:
Sq::StrArray is almost the same as Sq::StringArray of old version.

Definitions of SqStringArray and Sq::StringArray are keeped for compatibility with older versions.